→ Moves: The site info header card logic to Viewmodel slice
↑ Updates: the logic in MySiteViewModel to use
 SiteInfoHeaderCardViewModelSlice
